+++ packages/libs/e2e-helpers/src/types.ts
@@ -1,0 +1,5 @@
+export type Options = {
+  count?: number;
+  start?: number;
+  end?: number;
+};