+++ packages/sdks/nextjs-sdk/CHANGELOG.md
@@ -2,6 +2,41 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.14.6](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.5...nextjs-sdk-0.14.6) (2025-07-31)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.16.4`
+* `web-component` updated to version `3.44.3`
+## [0.14.5](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.4...nextjs-sdk-0.14.5) (2025-07-31)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.16.3`
+* `web-component` updated to version `3.44.2`
+## [0.14.4](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.3...nextjs-sdk-0.14.4) (2025-07-29)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.16.2`
+## [0.14.3](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.2...nextjs-sdk-0.14.3) (2025-07-27)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.16.1`
+* `web-component` updated to version `3.44.1`
+## [0.14.2](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.1...nextjs-sdk-0.14.2) (2025-07-22)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.16.0`
+* `web-component` updated to version `3.44.0`
+## [0.14.1](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.0...nextjs-sdk-0.14.1) (2025-07-21)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.15.0`
+* `web-component` updated to version `3.43.20`
 ## [0.14.0](https://github.com/descope/descope-js/compare/nextjs-sdk-0.13.20...nextjs-sdk-0.14.0) (2025-07-10)
 
 ### Dependency Updates